// typechain-types/TokenBridge.ts
export interface TokenBridgeInterface extends Interface {
  getFunction(nameOrSignature: "bridge" | "sourceToken"): FunctionFragment;

  getEvent(nameOrSignatureOrTopic: "Bridged"): EventFragment;

  encodeFunctionData(functionFragment: "bridge", values: [AddressLike]): string;
  encodeFunctionData(
    functionFragment: "sourceToken",
    values?: undefined
  ): string;

  decodeFunctionResult(functionFragment: "bridge", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "sourceToken",
    data: BytesLike
  ): Result;
}

export namespace BridgedEvent {
  export type InputTuple = [
    senderERC20Address: AddressLike,
    receiverBEP20Address: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [
    senderERC20Address: string,
    receiverBEP20Address: string,
    amount: bigint
  ];
  export interface OutputObject {
    senderERC20Address: string;
    receiverBEP20Address: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface TokenBridge extends BaseContract {
  connect(runner?: ContractRunner | null): TokenBridge;
  waitForDeployment(): Promise<this>;

  interface: TokenBridgeInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  bridge: TypedContractMethod<
    [receiverBEP20Address: AddressLike],
    [void],
    "nonpayable"
  >;

  sourceToken: TypedContractMethod<[], [string], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "bridge"
  ): TypedContractMethod<
    [receiverBEP20Address: AddressLike],
    [void],
    "nonpayable"
  >;
  getFunction(
    nameOrSignature: "sourceToken"
  ): TypedContractMethod<[], [string], "view">;

  getEvent(
    key: "Bridged"
  ): TypedContractEvent<
    BridgedEvent.InputTuple,
    BridgedEvent.OutputTuple,
    BridgedEvent.OutputObject
  >;

  filters: {
    "Bridged(address,address,uint256)": TypedContractEvent<
      BridgedEvent.InputTuple,
      BridgedEvent.OutputTuple,
      BridgedEvent.OutputObject
    >;
    Bridged: TypedContractEvent<
      BridgedEvent.InputTuple,
      BridgedEvent.OutputTuple,
      BridgedEvent.OutputObject
    >;
  };
}